Renamed files to Upper Camel Case and added a descriptive heade.
We can now query available GPUs from `nvidia-smi` and can compare them to the visible devices as seen by CUDA via nvcc
and TensorFlow. This is important as we are now able to verify that the number of GPUs requested via Slurm are all visible
by the application software we require them them for (i.e. programming in CUDA and/or Tensorflow). Should there
be a mismatch between the number of GPUs, we need to diagnose further before actually requiring multiple GPUs that
then would be idle in the application wasting resources.